feat: Convert the config file to TS
This patch converts lib/config to TS. The directory and util file are
combined into a single file, as its only about 5 lines of code, so the
split seems unnecessary.

`dotenv` and `camelcase` are also upgraded. Neither had breaking changes
that affected us (they were mostly dropping support for old Node
versions) and upgrading means we get type definitions.

Also removed the "silent" option from dotenv.config; this was removed in
v3.0.0 in 2016, but must've been missed. TS immediately highlighted that
this was not a supported option.
