+++ lib/auth-header.js
@@ -1,7 +1,7 @@
 const replace = require('./replace-vars');
 
 module.exports = ({ scheme, token, username, password }) => {
-  const config = require('./config');
+  const { config } = require('./config');
 
   if (scheme === 'token') {
     return `Token ${replace(token, config)}`;
