+++ lib/filters/index.js
@@ -27,7 +27,7 @@ const validateHeaders = (headerFilters, requestHeaders = []) => {
 // reads config that defines
 module.exports = (ruleSource) => {
   let rules = [];
-  const config = require('../config');
+  const { config } = require('../config');
 
   // polymorphic support
   if (Array.isArray(ruleSource)) {
