+++ lib/log.js
@@ -1,7 +1,7 @@
 const bunyan = require('bunyan');
 const escapeRegExp = require('lodash.escaperegexp');
 const mapValues = require('lodash.mapvalues');
-const config = require('./config');
+const { config } = require('./config');
 
 const sanitiseConfigVariable = (raw, variable) =>
   raw.replace(
