+++ package.json
@@ -45,10 +45,10 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "bunyan": "^1.8.12",
-    "camelcase": "^5.3.1",
+    "camelcase": "^6.0.0",
     "clarify": "^2.0.0",
     "clite": "^0.3.0",
-    "dotenv": "^6.2.0",
+    "dotenv": "^8.2.0",
     "ejson": "^2.2.0",
     "engine.io": "^3.4.0",
     "engine.io-client": "^3.4.0",
